VisualEncoding/requirements.txt
astroid==2.8.4
autopep8==1.6.0
cairocffi==1.0.0
CairoSVG==2.5.2
cffi==1.15.0
cssselect2==0.4.1
defusedxml==0.7.1
isort==5.10.1
lazy-object-proxy==1.6.0
lxml==4.6.5
mccabe==0.6.1
Pillow>=9.0.0
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.21
pylint==2.11.1
PyYAML==6.0
reportlab==3.6.2
svglib==1.1.0
svgutils==0.3.4
svgwrite==1.4.1
tinycss2==1.1.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
webencodings==0.5.1
wrapt==1.13.3
